+++ src/main/java/cn/banny/unidbg/arm/Arm64Hook.java
@@ -35,7 +35,7 @@ public abstract class Arm64Hook extends Arm64Svc {
     }
 
     @Override
-    public final int handle(Emulator emulator) {
+    public final long handle(Emulator emulator) {
         Unicorn u = emulator.getUnicorn();
         Pointer sp = UnicornPointer.register(emulator, Arm64Const.UC_ARM64_REG_SP);
         try {
@@ -43,7 +43,7 @@ public abstract class Arm64Hook extends Arm64Svc {
             sp = sp.share(-8);
             sp.setLong(0, status.jump);
 
-            return (int) status.returnValue;
+            return status.returnValue;
         } finally {
             u.reg_write(Arm64Const.UC_ARM64_REG_SP, ((UnicornPointer) sp).peer);
         }
