+++ src/main/java/cn/banny/unidbg/arm/ArmIntercept.java
@@ -45,7 +45,7 @@ public class ArmIntercept extends ArmSvc {
     }
 
     @Override
-    public int handle(Emulator emulator) {
+    public long handle(Emulator emulator) {
         Unicorn u = emulator.getUnicorn();
         if (callback != null) {
             callback.onIntercept(emulator);
