+++ src/main/java/cn/banny/unidbg/linux/ARMSyscallHandler.java
@@ -67,7 +67,7 @@ public class ARMSyscallHandler extends UnixSyscallHandler implements SyscallHand
             if (svcNumber != 0) {
                 Svc svc = svcMemory.getSvc(svcNumber);
                 if (svc != null) {
-                    u.reg_write(ArmConst.UC_ARM_REG_R0, svc.handle(emulator));
+                    u.reg_write(ArmConst.UC_ARM_REG_R0, (int) svc.handle(emulator));
                     return;
                 }
                 u.emu_stop();
