+++ src/main/java/cn/banny/unidbg/linux/android/ArmLD.java
@@ -42,14 +42,14 @@ public class ArmLD extends Dlfcn {
                 case "dlerror":
                     return svcMemory.registerSvc(new ArmSvc() {
                         @Override
-                        public int handle(Emulator emulator) {
-                            return (int) error.peer;
+                        public long handle(Emulator emulator) {
+                            return error.peer;
                         }
                     }).peer;
                 case "dlclose":
                     return svcMemory.registerSvc(new ArmSvc() {
                         @Override
-                        public int handle(Emulator emulator) {
+                        public long handle(Emulator emulator) {
                             long handle = ((Number) emulator.getUnicorn().reg_read(ArmConst.UC_ARM_REG_R0)).intValue() & 0xffffffffL;
                             if (log.isDebugEnabled()) {
                                 log.debug("dlclose handle=0x" + Long.toHexString(handle));
@@ -77,7 +77,7 @@ public class ArmLD extends Dlfcn {
                             }
                         }
                         @Override
-                        public int handle(Emulator emulator) {
+                        public long handle(Emulator emulator) {
                             Pointer filename = UnicornPointer.register(emulator, ArmConst.UC_ARM_REG_R0);
                             int flags = ((Number) emulator.getUnicorn().reg_read(ArmConst.UC_ARM_REG_R1)).intValue();
                             if (log.isDebugEnabled()) {
@@ -89,7 +89,7 @@ public class ArmLD extends Dlfcn {
                 case "dladdr":
                     return svcMemory.registerSvc(new ArmSvc() {
                         @Override
-                        public int handle(Emulator emulator) {
+                        public long handle(Emulator emulator) {
                             long addr = ((Number) emulator.getUnicorn().reg_read(ArmConst.UC_ARM_REG_R0)).intValue() & 0xffffffffL;
                             Pointer info = UnicornPointer.register(emulator, ArmConst.UC_ARM_REG_R1);
                             log.info("dladdr addr=0x" + Long.toHexString(addr) + ", info=" + info);
@@ -99,7 +99,7 @@ public class ArmLD extends Dlfcn {
                 case "dlsym":
                     return svcMemory.registerSvc(new ArmSvc() {
                         @Override
-                        public int handle(Emulator emulator) {
+                        public long handle(Emulator emulator) {
                             long handle = ((Number) emulator.getUnicorn().reg_read(ArmConst.UC_ARM_REG_R0)).intValue() & 0xffffffffL;
                             Pointer symbol = UnicornPointer.register(emulator, ArmConst.UC_ARM_REG_R1);
                             if (log.isDebugEnabled()) {
@@ -111,7 +111,7 @@ public class ArmLD extends Dlfcn {
                 case "dl_unwind_find_exidx":
                     return svcMemory.registerSvc(new ArmSvc() {
                         @Override
-                        public int handle(Emulator emulator) {
+                        public long handle(Emulator emulator) {
                             Pointer pc = UnicornPointer.register(emulator, ArmConst.UC_ARM_REG_R0);
                             Pointer pcount = UnicornPointer.register(emulator, ArmConst.UC_ARM_REG_R1);
                             if (log.isDebugEnabled()) {
@@ -125,7 +125,7 @@ public class ArmLD extends Dlfcn {
         return 0;
     }
 
-    private int dlopen(Memory memory, String filename, Emulator emulator) {
+    private long dlopen(Memory memory, String filename, Emulator emulator) {
         Pointer pointer = UnicornPointer.register(emulator, ArmConst.UC_ARM_REG_SP);
         try {
             Module module = memory.dlopen(filename, false);
@@ -161,7 +161,7 @@ public class ArmLD extends Dlfcn {
                     m.initFunctionList.clear();
                 }
 
-                return (int) module.base;
+                return module.base;
             }
         } catch (IOException e) {
             throw new IllegalStateException(e);
